Normal Distribution number generator

+++ src/job/tasks.c
@@ -6,7 +6,7 @@
 /*   By: kdaniely <kdaniely@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/29 15:31:01 by kdaniely          #+#    #+#             */
-/*   Updated: 2024/04/09 20:44:24 by kdaniely         ###   ########.fr       */
+/*   Updated: 2024/04/19 18:07:27 by kdaniely         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -91,7 +91,7 @@ static void	get_crops(t_darray *crops, t_image *r, int crop_c)
 	int		cut_count;
 
 	ft_darray_init(crops, sizeof(t_task), crop_c);
-	cut_count = floor(sqrt(crop_c));
+	cut_count = floor(sqrtf(crop_c));
 	delta.u = r->width / cut_count;
 	delta.v = r->height / cut_count;
 	if (delta.u < 1 && delta.v < 1)
